requirements.txt
Django==1.7.6
django-summernote==0.6.3
foursquare==2015.2.1
MySQL-python==1.2.5
Pillow==2.7.0
requests==2.6.0
six==1.9.0
django-bootstrap3-datetimepicker==2.2.3